requirements.txt
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.4
python-dotenv==1.0.1

google-api-python-client==2.137.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1

transformers==4.43.3
tokenizers==0.19.1
sentencepiece==0.2.0
accelerate==0.33.0
tabulate==0.9.0
